requirements.txt
fastapi
uvicorn
requests
chromadb
sentence-transformers
python-dotenv
